crawler/requirements.txt
# Python dependencies
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
